+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated to Intel Gaudi software Release 1.17.0 ([#221](https://github.com/Lightning-AI/lightning-Habana/pull/221))
 - Modified torch device specification for FSDP on HPU ([#222](https://github.com/Lightning-AI/lightning-Habana/pull/222))
 - Updated strategy to use default fork ([#234](https://github.com/Lightning-AI/lightning-Habana/pull/234))
+- Generalized hpu parallel strategies ([#237](https://github.com/Lightning-AI/lightning-Habana/pull/237))
 
 ### Fixed